moved parser and document functions to modules print and parse
