+++ src/config.rs
@@ -1,18 +1,22 @@
 //! Config objects, to be read from Yarner.toml
 
-use crate::parser::md::MdParser;
 use crate::util::Fallible;
-use serde::{Deserialize, Serialize};
+use once_cell::sync::Lazy;
+use regex::Regex;
+use serde::{de::Error as _, Deserialize, Deserializer, Serialize};
 use std::collections::HashMap;
 use std::fs::read_to_string;
 use std::path::Path;
 use toml::from_str;
 
+const LINK_PATTERN: &str = r"\[([^\[\]]*)\]\((.*?)\)";
+pub static LINK_REGEX: Lazy<Regex> = Lazy::new(|| Regex::new(LINK_PATTERN).unwrap());
+
 /// Top-level config
 #[derive(Deserialize, Debug)]
 pub struct Config {
     /// Config for Markdown parser
-    pub parser: MdParser,
+    pub parser: ParserSettings,
     /// Config for paths
     #[serde(default)]
     pub paths: Paths,
@@ -38,6 +42,81 @@ impl Config {
 
         Ok(())
     }
+}
+
+/// The config for parsing a Markdown document
+#[derive(Clone, Deserialize, Debug)]
+pub struct ParserSettings {
+    /// The sequence that identifies the start and end of a fenced code block
+    pub fence_sequence: String,
+    /// Alternative sequence that identifies the start and end of a fenced code block.
+    /// Allows for normal Markdown fences in code blocks
+    pub fence_sequence_alt: String,
+    /// The language to set if there was no automatically detected language. Optional
+    pub default_language: Option<String>,
+    /// Temporary switch to disable comment extraction
+    #[serde(default)]
+    pub comments_as_aside: bool,
+    /// The sequence to identify a comment which should be omitted from the compiled code, and may
+    /// be rendered as an `<aside>` if `comments_as_aside` is set.
+    pub block_name_prefix: String,
+    /// The sequence to identify the start of a macro invocation.
+    pub macro_start: String,
+    /// The sequence to identify the end of a macro invocation.
+    pub macro_end: String,
+    /// The sequence to identify the start of a transclusion.
+    pub transclusion_start: String,
+    /// The sequence to identify the end of a transclusion.
+    pub transclusion_end: String,
+    /// Prefix for links that should be followed during processing.
+    /// Should be RegEx-compatible.
+    #[serde(rename(deserialize = "link_prefix"))]
+    #[serde(deserialize_with = "from_link_prefix")]
+    pub link_following_pattern: (String, Regex),
+    /// Prefix for file-specific entry points.
+    pub file_prefix: String,
+    /// Name prefix for code blocks not shown in the docs.
+    pub hidden_prefix: String,
+}
+
+impl ParserSettings {
+    pub fn check(&self) -> Result<(), String> {
+        if self.comments_as_aside {
+            Err(r#"Comment extraction is temporarily disabled.
+Please comment out option `comments_as_aside` until the next version, and rename `comment_start` to `block_name_prefix`"#.to_string())
+        } else {
+            Ok(())
+        }
+    }
+
+    /// Sets the default language of the returned parser (or does nothing if `None` is passed)
+    pub fn default_language(&self, language: Option<String>) -> Self {
+        let mut cloned = self.clone();
+
+        if language.is_some() {
+            cloned.default_language = language;
+        }
+
+        cloned
+    }
+}
+
+fn from_link_prefix<'de, D>(deserializer: D) -> Result<(String, Regex), D::Error>
+where
+    D: Deserializer<'de>,
+{
+    let prefix: &str = Deserialize::deserialize(deserializer)?;
+    Ok((
+        prefix.to_string(),
+        Regex::new(&format!("{}{}", prefix, LINK_PATTERN)).map_err(|err| {
+            D::Error::custom(format!(
+                "Error compiling Regex pattern {}{}\n{}",
+                prefix,
+                LINK_PATTERN,
+                err.to_string()
+            ))
+        })?,
+    ))
 }
 
 /// Config for paths
