+++ src/parser/mod.rs
@@ -1,2 +1,1 @@
 pub mod code;
-pub mod md;
